962 upgrade (#305)
* 962 upgrade

Fixes https://github.com/nemtech/symbol-sdk-java/issues/307
Fixes https://github.com/nemtech/symbol-sdk-java/issues/304

+++ sdk-core/src/main/java/io/nem/symbol/sdk/model/account/ActivityBucket.java
@@ -36,7 +36,7 @@ public class ActivityBucket {
     /**
      * Beneficiary count.
      */
-    private final int beneficiaryCount;
+    private final long beneficiaryCount;
 
     /**
      * Raw score.
@@ -52,7 +52,7 @@ public class ActivityBucket {
      * @param beneficiaryCount Beneficiary count.
      * @param rawScore Raw score.
      */
-    public ActivityBucket(BigInteger startHeight, BigInteger totalFeesPaid, int beneficiaryCount,
+    public ActivityBucket(BigInteger startHeight, BigInteger totalFeesPaid, long beneficiaryCount,
         BigInteger rawScore) {
         this.startHeight = startHeight;
         this.totalFeesPaid = totalFeesPaid;
@@ -68,7 +68,7 @@ public class ActivityBucket {
         return totalFeesPaid;
     }
 
-    public int getBeneficiaryCount() {
+    public long getBeneficiaryCount() {
         return beneficiaryCount;
     }
