+++ sdk-vertx-client/src/main/java/io/nem/symbol/sdk/infrastructure/vertx/mappers/MosaicMetadataTransactionMapper.java
@@ -20,7 +20,6 @@ import static io.nem.symbol.core.utils.MapperUtils.toUnresolvedAddress;
 
 import io.nem.symbol.core.utils.ConvertUtils;
 import io.nem.symbol.core.utils.MapperUtils;
-import io.nem.symbol.sdk.model.account.PublicAccount;
 import io.nem.symbol.sdk.model.account.UnresolvedAddress;
 import io.nem.symbol.sdk.model.mosaic.UnresolvedMosaicId;
 import io.nem.symbol.sdk.model.network.NetworkType;
@@ -58,7 +57,7 @@ class MosaicMetadataTransactionMapper extends
             scopedMetaDataKey,
             value);
         factory.valueSizeDelta(valueSizeDelta);
-        Integer valueSize = transaction.getValueSize();
+        Long valueSize = transaction.getValueSize();
         if (valueSize != null) {
             factory.valueSize(valueSize);
         }
